Warning: this revision may cause your computer to burst into flames. OK, probably not, but it's not recommended for actual use.
Worked most of the freeglut render funcs (glutSolidSphere(), etc.) into
szg & removed most of the GLUT dependencies.

Issues:
Only tested build on win32/MinGW.
No support for glutBitmapXXXX(). In the szg core, this is used in
arFramerateGraph (now commented out) and in vrtest, which will need to be
modified (or support these funcs, whichever's most convenient).
There are still a couple of utilities that use GLUT for windowing:
szg-rp and DeskCalibrator, neither of which I've _ever_ used. Update them
if it's easy, else make the overall build ignore errors from building them.